added

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: nnasiri <nnasiri@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 16:13:49 by nnasiri           #+#    #+#             */
-/*   Updated: 2024/04/09 16:17:01 by nnasiri          ###   ########.fr       */
+/*   Updated: 2024/04/16 14:37:20 by nnasiri          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
